+++ types/configs/editor-config.d.ts
@@ -6,7 +6,7 @@ import {I18nConfig} from './i18n-config';
 export interface EditorConfig {
   /**
    * Element where Editor will be append
-   * @deprecated property will be removed in next major release, use holder instead
+   * @deprecated property will be removed in the next major release, use holder instead
    */
   holderId?: string | HTMLElement;